Get the treatment selectors to start working
